[MINOR] Refactor component tests (move layout tests into componentx)
