+++ src/test/java/org/apache/sysds/test/componentx/compress/JolEstimateTest.java
@@ -1,4 +1,4 @@
-package org.apache.sysds.test.component.compress.colgroup;
+package org.apache.sysds.test.componentx.compress;
 
 import static org.junit.Assert.assertTrue;
